+++ src/apis/auth/discord.py
@@ -151,7 +151,7 @@ async def get_callback(request: Request, response: Response, code: Optional[str]
                 discord_embed = {"title": ml.tr(request, "new_login_title", force_lang = language),
                                  "description": "",
                                  "fields": [{"name": ml.tr(request, "country", force_lang = language), "value": getRequestCountry(request), "inline": True},
-                                            {"name": ml.tr(request, "ip", force_lang = language), "value": request.client.host, "inline": True}]
+                                            {"name": ml.tr(request, "ip", force_lang = language), "value": f"`{request.client.host}`", "inline": True}]
                 }
             )